flatten logger
