+++ packages/insomnia-inso/src/commands/export-specification.test.ts
@@ -1,8 +1,8 @@
 import { afterEach, beforeAll, beforeEach, describe, expect, it, jest } from '@jest/globals';
 import { MockedFunction } from 'jest-mock';
 
+import { logger } from '../cli';
 import { globalBeforeAll, globalBeforeEach } from '../jest/before';
-import { logger } from '../logger';
 import { writeFileWithCliOptions as _writeFileWithCliOptions } from '../write-file';
 import { exportSpecification } from './export-specification';
