+++ packages/insomnia-inso/src/commands/run-tests.ts
@@ -1,11 +1,10 @@
 import { generate, runTestsCli, TestSuite } from 'insomnia-testing';
 
+import { type GlobalOptions, logger, noConsoleLog } from '../cli';
 import { loadDb } from '../db';
 import { loadEnvironment, promptEnvironment } from '../db/models/environment';
 import type { UnitTest, UnitTestSuite } from '../db/models/types';
 import { loadTestSuites, promptTestSuites } from '../db/models/unit-test-suite';
-import type { GlobalOptions } from '../get-options';
-import { logger, noConsoleLog } from '../logger';
 
 export type TestReporter = 'dot' | 'list' | 'spec' | 'min' | 'progress';
